requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0
python-multipart==0.0.6
